backend/pyproject.toml
[tool.poetry]
name = "chat-data-backend"
version = "0.1.0"
description = "AI-powered data analysis platform"
authors = ["kischn"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
openai = "^1.12.0"
pandas = "^2.1.4"
celery = "^5.3.4"
redis = "^5.0.1"
minio = "^7.2.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
ruff = "^0.1.6"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
